Split out atomic info functions into their own source file.
Create a Info class as a subclass of Atomic and move all of the
functions and CLI to this file.
